+++ test/runtests.jl
@@ -13,4 +13,5 @@ using DroopOPF
     include("test_matpower.jl")
     include("test_plotting.jl")
     include("test_droop_sweep.jl")
+    include("test_droop_optimization.jl")
 end